Add another hook into components lifecycle
When we navigate to the Graphs view comming from other view
then the entries are already loaded and so there's no re-rendeding
of the Graphs view (which in turn re-renders the Graph views).

As the Graph views are not re-rendered their
'componentWillReceiveProps' is never called.

To fix it we use also the 'componentDidMount' hook.